refactor: 💡 sepearte hooks and componet responsibiltiy
introduce hook for clean code instead of managing everything in
component
