+++ application-frontend/src/app/(app)/upload/page.js
@@ -1,8 +1,7 @@
 import Header from '@/app/(app)/Header'
-import CsvFileInput from '@/components/CsvFileInput'
-
+import KeywordCsvFileInput from '@/components/CsvFileInput'
 export const metadata = {
-    title: 'Laravel - Dashboard',
+    title: 'Web Crawler - Dashboard',
 }
 
 const Upload = () => {
@@ -13,7 +12,7 @@ const Upload = () => {
                 <div className="max-w-7xl mx-auto sm:px-6 lg:px-8">
                     <div className="bg-white overflow-hidden shadow-sm sm:rounded-lg mx-2 md:mx-auto lg:mx-auto">
                         <div className="p-6 bg-white border-b border-gray-200">
-                            <CsvFileInput />
+                            <KeywordCsvFileInput />
                         </div>
                     </div>
                 </div>
